feat: 올린 장소 서버로부터 불러오는 로직 커스텀 훅으로 분리
